카드 소유자 이름 부분을 컴포넌트로 분리
